\name{paeonia}
\alias{paeonia}
\title{phenotypic integration example dataset #1}
\description{phenotypic integration example dataset from Méndez & Traveset (2003).}
\usage{data(paeonia)}
\details{
This data set represents the dry mass (in grams) of inflorescence components of the herbaceous perennial plant \emph{Paeonia cambessedesii}. Variables include mass for 'PETIOLE',	'RECEPTACLE',	'SEPAL', 'LEAF_ACC',	'PETAL',	'STAMEN', and	'CARPEL'. 'LEAF_ACC' represent the dry mass of the accessory leaf (subtending the flower but close to the sepals).	Furthermore the first and last columns include the individual number ('ID') and the total weight of the inflorescence ('TOTAL').}
\references{Méndez, M. & Traveset, A. 2003. Sexual allocation in single-flowered individuals of a hermaphroditic species in relation to plant and flower size. Oecologia 137:69-75.}
\author{M. Méndez, A. Traveset}
\examples{
data(paeonia)
}
