\name{PHENIX-package}
\alias{PHENIX-package}
\alias{PHENIX-package}
\docType{package}
\title{size-controlled phenotypic integration index}
\description{PHENIX estimates a phenotypic integration index taking into account the size of the studied organ. For that, the magnitude of phenotypic integration between a set of traits in a population is quantified by the variance of the eigenvalues of the partial-correlation matrix (Torices & Méndez 2014). This partial-correlation matrix assess the correlation structure between all traits after controlling by the organ size. In addition, PHENIX also computes the integration values based on the correlation matrix between traits (Wagner 1984). Confidence intervals of both indices are calculated by bootstrapping and jackknife and the statistical significance is calculated by simulating a null distribution.
}
\details{
\tabular{ll}{
Package: \tab PHENIX\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-1-22\cr
License: \tab GPL-2\cr
}
}
\author{R. Torices, A.J. Muñoz-Pajares

Maintainer: A.J. Muñoz-Pajares to <ajesusmp@cibio.up.pt>
}

\references{
R Torices, AJ Muñoz-Pajares (2015). PHENIX: An R package to estimate a size-controlled phenotypic integration index Applications in plant sciences 3 (5) DOI: 10.3732/apps.1400104
}

