\name{tussilago}
\alias{tussilago}
\title{phenotypic integration example dataset #2}
\description{phenotypic integration example dataset from Torices & Méndez (2014).}
\usage{data(tussilago)}
\details{This data set represents the dry mass (in grams) of inflorescence components of the sunflower species \emph{Tussilago farfara}. The inflorescences were dissected in 'SCAPE', 'RECEPTACLE', 'MALEFL' (male flowers), 'OVAR' (reproductive part of female flowers), and 'RAYS' (the petaloid ray of female flower). Furthermore in the last column the total weight of the inflorescence is added 'TOTALSIZE'.}
\references{Torices, R. & Méndez, M. 2014. Resource allocation to inflorescence components is highly integrated despite differences between allocation currencies and sites. International Journal of Plant Science 175(6):713–723.}
\author{R. Torices, M. Méndez}
\examples{
data(tussilago)
}
