\name{PHENIX-package}
\alias{PHENIX-package}
\alias{PHENIX-package}
\docType{package}
\title{size-controlled phenotypic integration index}
\description{PHENIX estimates a phenotypic integration index taking into account the size of the studied organ. For that, the magnitude of phenotypic integration between a set of traits in a population is quantified by the variance of the eigenvalues of the partial-correlation matrix (Torices & Méndez 2014). This partial-correlation matrix assess the correlation structure between all traits after controlling by the organ size. In addition, PHENIX also computes the integration values based on the correlation matrix between traits (Wagner 1984). Statistical significance of both indices is assessed by bootstrapping.}
\details{
\tabular{ll}{
Package: \tab PHENIX\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-14\cr
License: \tab GPL-2\cr
}
}
\author{R. Torices, A. J. Muñoz-Pajares

Maintainer: A. J. Muñoz-Pajares to <ajesusmp@cibio.up.pt>

}

